Build: Renaming package to project name
