+++ mobile/app/src/androidTest/java/nom4d/toysrevive/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package nom4d.myapplication
+package nom4d.toysrevive
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
@@ -17,6 +17,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("nom4d.myapplication", appContext.packageName)
+        assertEquals("nom4d.toysrevive", appContext.packageName)
     }
 }
