+++ mobile/app/src/main/java/nom4d/toysrevive/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package nom4d.myapplication.ui.theme
+package nom4d.toysrevive.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
